meetup_assistants/requirements.txt
#AI-Agents
duckduckgo-search==5.3.0b4
litellm==1.40.22
agentops==0.2.6
apscheduler==3.10.4
# X Account
tweepy==4.14.0

#AutoGen
pyautogen==0.2.32
pyautogen[groq]==0.2.32


#FastAPI
fastapi==0.111.0
pydantic==2.7.4
uvicorn==0.30.1
#apscheduler==3.10.4

#Postgres
psycopg==3.1.19
